separate templates from specializations
